Move checked_(re|c)alloc to rd_util.hpp
